01_RASPBERRY_PI_ZERO_2_WH/requirements.txt
paho-mqtt==1.6.1
requests==2.31.0
pygame==2.5.2
python-dotenv==1.0.0
